+++ CHANGELOG.md
@@ -5,12 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-# Unreleased
+# [5.3.0] - 2025-08-07
 
 ### Added
 
 - Added `Content.simplify` https://github.com/Textualize/textual/pull/6023
 - Added `textual.reactive.Initialize` https://github.com/Textualize/textual/pull/6023
+
+### Fixed
+
+- Fixed issue with IDs in markdown https://github.com/Textualize/textual/pull/6019 https://github.com/Textualize/textual/pull/6023
 
 ## [5.2.0] - 2025-08-01
 
@@ -3054,6 +3058,7 @@ https://textual.textualize.io/blog/2022/11/08/version-040/#version-040
 - New handler system for messages that doesn't require inheritance
 - Improved traceback handling
 
+[5.3.0]: https://github.com/Textualize/textual/compare/v5.2.0...v5.3.0
 [5.2.0]: https://github.com/Textualize/textual/compare/v5.1.1...v5.2.0
 [5.1.1]: https://github.com/Textualize/textual/compare/v5.1.0...v5.1.1
 [5.1.0]: https://github.com/Textualize/textual/compare/v5.0.1...v5.1.0